New Distance Algorithm
New distance algorithm used which has accuracy up to 100in in front of
the target.
The angle to horizontal for the camera must be set to 20 degrees and the
view angle to 30.7 degrees.
Also, you can use the switch on filter for the image on the dashboard.
Two batch files in the ip change folder: a changes to 10.40.1.9 and b
changes back to automatic ip.