Distance Calibration
Today I realized that my calculations from yesterday were bogus!
I plotted true distance against the distance that our tangent code was
giving and it gave a linear relationship (can be found in excel file
under IP address change folder)
The dashboard is currently coded to use the linear relationship (y=mx+b)
using the parameters: parameter 1 = m, and parameter 2 = b.
The values that are in the dashboard main vi are the ones found in excel
but they may need to be changed a little.
Hence, I don't think we need any new algorithms to make the distance
work, this is accurate enough!
Also the yellow lights work, if anyone was wondering...