Updated autonomous code for our new state machine
Moved hot goal detection to occur right after the autonomous mode
starts. This code is only "silver" as its not completely working.. the
state machine works fine when the "Hot or Not" state spits out "Hot" but
there is an issue with the "Not" condition for now which needs to be
debugged.